requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
deepl==1.17.0
distro==1.9.0
exceptiongroup==1.2.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
mutagen==1.47.0
numpy==1.26.4
openai==1.10.0
packaging==24.0
pandas==2.2.2
pathlib==1.0.1
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.1.0
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
st-annotated-text==4.0.1
streamlit==1.33.0
streamlit-ext==0.1.10
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
youtube-transcript-api==0.6.2
yt-dlp==2024.4.9
